requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
aiohttp>=3.8.0
openai>=1.0.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
pyyaml==6.0.1
tqdm>=4.65.0
outlines>=0.0.1
chromadb>=0.4.18 